+++ experimental/tools/unit-generators/vhdl/generators/handshake/merge.py
@@ -1,6 +1,6 @@
 from generators.support.utils import VhdlScalarType
 from generators.support.merge_notehb import generate_merge_notehb
-from generators.handshake.tehb import generate_tehb
+from generators.support.tehb import generate_tehb
 
 
 def generate_merge(name, params):
