+++ experimental/tools/unit-generators/vhdl/generators/support/ofifo.py
@@ -1,5 +1,5 @@
 from generators.support.utils import VhdlScalarType
-from generators.handshake.tehb import generate_tehb
+from generators.support.tehb import generate_tehb
 from generators.support.elastic_fifo_inner import generate_elastic_fifo_inner
